+++ package.json
@@ -18,6 +18,7 @@
     "@babel/preset-env": "7.8.3",
     "@babel/preset-react": "7.8.3",
     "@types/jest": "25.1.4",
+    "@types/leaflet": "1.5.11",
     "@types/react": "16.9.23",
     "@types/react-test-renderer": "16.9.2",
     "@typescript-eslint/eslint-plugin": "2.24.0",
@@ -31,18 +32,16 @@
     "eslint-config-htmlacademy": "0.5.2",
     "eslint-plugin-react": "7.18.0",
     "jest": "25.1.0",
-    "python": "0.0.4",
     "react-test-renderer": "16.12.0",
     "redux-devtools-extension": "2.13.8",
     "redux-mock-store": "1.5.4",
     "ts-jest": "25.2.1",
     "ts-loader": "6.2.1",
-    "tslint": "6.1.0",
     "typescript": "3.8.3",
     "utility-types": "3.10.0",
-    "webpack": "4.41.5",
-    "webpack-cli": "3.3.10",
-    "webpack-dev-server": "3.10.1"
+    "webpack": "4.42.0",
+    "webpack-cli": "3.3.11",
+    "webpack-dev-server": "3.10.3"
   },
   "engines": {
     "node": "10"
@@ -52,14 +51,13 @@
     "convert-keys": "1.3.4",
     "leaflet": "1.6.0",
     "moment": "2.24.0",
-    "prop-types": "15.7.2",
+    "npm": "6.14.3",
     "react": "16.12.0",
     "react-dom": "16.12.0",
     "react-redux": "7.2.0",
     "react-router-dom": "5.1.2",
     "redux": "4.0.5",
     "redux-thunk": "2.3.0",
-    "reselect": "4.0.0",
-    "tslint-react": "^4.2.0"
+    "reselect": "4.0.0"
   }
 }
