+++ src/components/city-nav-item/city-nav-item.test.tsx
@@ -1,15 +1,16 @@
-import React from 'react';
-import renderer from 'react-test-renderer';
-import {MemoizedCityNavItem} from "./city-nav-item.tsx";
+import * as React from 'react';
+import * as renderer from 'react-test-renderer';
+import {MemoizedCityNavItem} from "./city-nav-item";
 
 const CITY_ID = 0;
 
 it(`Render MemoizedCityNavItem`, () => {
+  const dispatch = jest.fn();
   const tree = renderer
     .create(
         <MemoizedCityNavItem
           city = {`Paris`}
-          setSelectedCity = {() => {}}
+          setSelectedCity = {dispatch}
           cityId = {CITY_ID}
           selectedCityId = {CITY_ID}
         />
