+++ src/components/favorites/favorites.test.tsx
@@ -1,15 +1,23 @@
-import React from 'react';
-import renderer from 'react-test-renderer';
-import Main from "./main.tsx";
-import {Provider} from "react-redux";
-import configureStore from "redux-mock-store";
-import {InitValue} from "../../reducers/cards-sorting-menu/cards-sorting-menu";
+import * as React from 'react';
+import * as renderer from 'react-test-renderer';
+import Favorites from "./favorites";
+import configureMockStore from "redux-mock-store";
 import {NameSpace} from "../../reducers/name-space";
+import {Provider} from "react-redux";
 import {BrowserRouter} from "react-router-dom";
+import MockAdapter from "axios-mock-adapter";
+import createAPI from "../../api";
+import thunk from 'redux-thunk';
+import {CardType} from "../../types";
 
-const mockStore = configureStore([]);
+const testFn = jest.fn();
 
-const cards = [
+const api = createAPI(testFn);
+
+const middlewares = [thunk.withExtraArgument(api)];
+const mockStore = configureMockStore(middlewares);
+
+const cards: CardType[] = [
   {
     id: 0,
     city: {
@@ -80,42 +88,20 @@ const cards = [
   }
 ];
 
-const reviews = [
-  {
-    id: 0,
-    comment: `A quiet cozy and picturesque that hides behind a a river by the unique lightness of Amsterdam.
-          The building is green and from 18th century.`,
-    date: new Date(2020, 1, 14, 0, 0, 0, 0).toISOString(),
-    rating: 4.5,
-    user: {
-      id: 0,
-      name: `Max`,
-      avatarUrl: `/img/avatar-max.jpg`,
-      isPro: true
-    },
-  },
-  {
-    id: 1,
-    comment: `A quiet cozy and picturesque that hides behind a a river by the unique lightness of Amsterdam.
-          The building is green and from 18th century.`,
-    date: new Date(2020, 0, 1, 0, 0, 0, 0).toISOString(),
-    rating: 3.5,
-    user: {
-      id: 1,
-      name: `Bob`,
-      avatarUrl: `/img/avatar.svg`,
-      isPro: true
-    },
-  }
-];
+it(`Render Favorites`, () => {
+  const apiMock = new MockAdapter(api);
 
-it(`Render Main`, () => {
+  apiMock
+    .onGet(`/favorite`)
+    .reply(200, cards);
+
+
   const store = mockStore({
-    [NameSpace.CARDS]: {
-      cards
+    [NameSpace.FAVORITES]: {
+      favorites: cards,
     },
     [NameSpace.USER]: {
-      userAuthorized: false,
+      userAuthorized: true,
       userInfo: {
         id: 1,
         name: `Oliver.conner`,
@@ -124,31 +110,15 @@ it(`Render Main`, () => {
         isPro: false
       },
     },
-    [NameSpace.CITY_SELECT]: {
-      cityName: 0
-    },
-    [NameSpace.CARD_SELECT]: {
-      cardId: 0
-    },
-    [NameSpace.CARDS_SORTING_MENU]: {
-      sortingMethodId: InitValue.INITIAL_SORTING_METHOD_ID,
-      menuState: InitValue.INITIAL_MENU_STATE
-    },
-    [NameSpace.REVIEWS]: {
-      reviews
-    },
-    [NameSpace.SIMILAR_OFFERS]: {
-      similarOffers: [cards[1]]
-    },
   });
 
   const tree = renderer
     .create(
-        <Provider store={store}>
-          <BrowserRouter>
-            <Main/>
-          </BrowserRouter>
-        </Provider>
+        <BrowserRouter>
+          <Provider store={store}>
+            <Favorites />
+          </Provider>
+        </BrowserRouter>
     )
     .toJSON();
   expect(tree).toMatchSnapshot();
