+++ src/components/header/header.test.tsx
@@ -1,20 +1,16 @@
-import React from 'react';
-import renderer from 'react-test-renderer';
-import Favorites from "./favorites.tsx";
-import configureMockStore from "redux-mock-store";
-import {NameSpace} from "../../reducers/name-space";
+import * as React from 'react';
+import * as renderer from 'react-test-renderer';
+import Header from "./header";
 import {Provider} from "react-redux";
+import configureStore from "redux-mock-store";
+import {InitValue} from "../../reducers/cards-sorting-menu/cards-sorting-menu";
+import {NameSpace} from "../../reducers/name-space";
 import {BrowserRouter} from "react-router-dom";
-import MockAdapter from "axios-mock-adapter";
-import createAPI from "../../api";
-import thunk from 'redux-thunk';
+import {CardType, ReviewType} from "../../types";
 
-const api = createAPI(() => {});
+const mockStore = configureStore([]);
 
-const middlewares = [thunk.withExtraArgument(api)];
-const mockStore = configureMockStore(middlewares);
-
-const cards = [
+const cards: CardType[] = [
   {
     id: 0,
     city: {
@@ -85,20 +81,42 @@ const cards = [
   }
 ];
 
-it(`Render Favorites`, () => {
-  const apiMock = new MockAdapter(api);
+const reviews: ReviewType[] = [
+  {
+    id: 0,
+    comment: `A quiet cozy and picturesque that hides behind a a river by the unique lightness of Amsterdam.
+          The building is green and from 18th century.`,
+    date: new Date(2020, 1, 14, 0, 0, 0, 0).toISOString(),
+    rating: 4.5,
+    user: {
+      id: 0,
+      name: `Max`,
+      avatarUrl: `/img/avatar-max.jpg`,
+      isPro: true
+    },
+  },
+  {
+    id: 1,
+    comment: `A quiet cozy and picturesque that hides behind a a river by the unique lightness of Amsterdam.
+          The building is green and from 18th century.`,
+    date: new Date(2020, 0, 1, 0, 0, 0, 0).toISOString(),
+    rating: 3.5,
+    user: {
+      id: 1,
+      name: `Bob`,
+      avatarUrl: `/img/avatar.svg`,
+      isPro: true
+    },
+  }
+];
 
-  apiMock
-    .onGet(`/favorite`)
-    .reply(200, cards);
-
-
+it(`Render Header`, () => {
   const store = mockStore({
-    [NameSpace.FAVORITES]: {
-      favorites: cards,
+    [NameSpace.CARDS]: {
+      cards
     },
     [NameSpace.USER]: {
-      userAuthorized: true,
+      userAuthorized: false,
       userInfo: {
         id: 1,
         name: `Oliver.conner`,
@@ -107,15 +125,31 @@ it(`Render Favorites`, () => {
         isPro: false
       },
     },
+    [NameSpace.CITY_SELECT]: {
+      cityName: 0
+    },
+    [NameSpace.CARD_SELECT]: {
+      cardId: 0
+    },
+    [NameSpace.CARDS_SORTING_MENU]: {
+      sortingMethodId: InitValue.INITIAL_SORTING_METHOD_ID,
+      menuState: InitValue.INITIAL_MENU_STATE
+    },
+    [NameSpace.REVIEWS]: {
+      reviews
+    },
+    [NameSpace.SIMILAR_OFFERS]: {
+      similarOffers: [cards[1]]
+    },
   });
 
   const tree = renderer
     .create(
-        <BrowserRouter>
-          <Provider store={store}>
-            <Favorites />
-          </Provider>
-        </BrowserRouter>
+        <Provider store={store}>
+          <BrowserRouter>
+            <Header />
+          </BrowserRouter>
+        </Provider>
     )
     .toJSON();
   expect(tree).toMatchSnapshot();
